[packaging][yum] use only datadog repo
Modified yum install line to disable all other repos and only enable the
DataDog repo. This fixes situations where a system might have broken repos, or
repos that aren't available in specific environments. An example would be a
company with an internal repo which is available during a build phase, but not
available in a prod environment.

Rebased: [quentin.madec@datadoghq.com] quote 'datadog'